split chad
