+++ sd_webui_bayesian_merger/optimiser.py
@@ -58,11 +58,11 @@ class BayesianOptimiser:
         self.iteration += 1
 
         if self.iteration == 1:
-            print('\n'+'-'*10+' warmup '+'-'*10+'>')
+            print("\n" + "-" * 10 + " warmup " + "-" * 10 + ">")
         elif self.iteration == self.init_points + 1:
-            print('\n'+'-'*10+' optimisation '+'-'*10+'>')
+            print("\n" + "-" * 10 + " optimisation " + "-" * 10 + ">")
 
-        it_type = 'warmup' if self.iteration <= self.init_points else 'optimisation'
+        it_type = "warmup" if self.iteration <= self.init_points else "optimisation"
         print(f"\n{it_type} - Iteration: {self.iteration}")
 
         weights = [params[f"block_{i}"] for i in range(25)]
@@ -115,7 +115,7 @@ class BayesianOptimiser:
         )
 
         # clean up and remove the last merge
-        self.merger.remove_previous_ckpt(self.iteration+1)
+        self.merger.remove_previous_ckpt(self.iteration + 1)
 
     def postprocess(self) -> None:
         for i, res in enumerate(self.optimizer.res):
